investment-scout/requirements.txt
# Investment Scout Requirements

# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Google APIs (Gmail)
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0

# AI/LLM
openai>=1.10.0

# Research (optional - install as needed)
# exa-py>=1.0.0
# tavily-python>=0.3.0

# PDF Processing (for deck analysis)
pymupdf>=1.23.0

# Utilities
asyncio-throttle>=1.0.0
